+++ src/ActivityLoggerServiceProvider.php
@@ -1,21 +1,11 @@
-<?php namespace JeroenG\ActivityLogger;
+<?php
+
+namespace JeroenG\ActivityLogger;
 
 use Illuminate\Support\ServiceProvider;
 
-/**
- * This is the service provider for Laravel.
- *
- * Place the line below in the providers array inside config/app.php
- * <code>JeroenG\ActivityLogger\ActivityLoggerServiceProvider::class,</code>
- * And this line into the alias array in the same file
- * <code>'Activity' => JeroenG\ActivityLogger\Facades\ActivityLogger::class,</code>
- * 
- * @package ActivityLogger
- * @author 	JeroenG
- * 
- **/
-class ActivityLoggerServiceProvider extends ServiceProvider {
-
+class ActivityLoggerServiceProvider extends ServiceProvider
+{
 	/**
 	 * Indicates if loading of the provider is deferred.
 	 *
@@ -30,6 +20,8 @@ class ActivityLoggerServiceProvider extends ServiceProvider {
 	 */
 	public function boot()
 	{
+        $this->loadMigrationsFrom( __DIR__.'/../migrations');
+
 		$this->publishes([
             __DIR__.'/../migrations' => $this->app->databasePath().'/migrations',
         ]);
@@ -42,11 +34,9 @@ class ActivityLoggerServiceProvider extends ServiceProvider {
 	 */
 	public function register()
 	{
-		$this->app['activitylogger'] = $this->app->share( function ($app)
-            {
+		$this->app->singleton('activitylogger', function ($app) {
                 return new ActivityLogger;
-            }
-        );
+		});
 	}
 
 	/**
@@ -56,7 +46,6 @@ class ActivityLoggerServiceProvider extends ServiceProvider {
 	 */
 	public function provides()
 	{
-		return array('activitylogger');
+		return ['activitylogger'];
 	}
-
 }